requirements.txt
langchain
langchain_community
langchain-chroma
langchain-groq
langchain-ollama
numpy
pandas
langserve
fastapi
streamlit
langsmith
python-dotenv
-e .
